pyproject.toml
[project]
name = "data-transformation-framework"
version = "0.1.0"
description = "A Python framework inspired by dbt and SQLMesh for SQL transformations on Snowflake"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Data Team"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Database",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "snowflake-connector-python>=3.6.0",
    "click>=8.1.0",
    "watchdog>=3.0.0",
    "PyYAML>=6.0",
    "sqlglot>=20.0.0",
    "Jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "colorama>=0.4.6",
    "tabulate>=0.9.0",
    "typing-extensions>=4.8.0",
]

[project.scripts]
framework = "framework.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

